Fix input labels attribute and update README
Updated `for` attribute on labels in the form component to correctly match associated input fields, improving accessibility. The changes were made to the `app.component.html` file. The `for` attribute was modified from "name" to "firstName" and "lastName" respectively for consistency and to prevent potential confusion.

Additionally, updated `README.md` to better describe the purpose and functionality of the project. The project title was changed and an illustrative code block was added to demonstrate how a form is structured in the component. Some irrelevant sections were also removed in order to enhance readability. The changes provide more clarity and context to users or different stakeholders who may refer to the README for understanding the project.