Remove unused variable and inject FormBuilder in AppComponent
In the `app.component.ts` file, the unused `title` variable was removed from the AppComponent to declutter the code and maintain cleaner codebase.

Also, instead of complete import, `FormBuilder` was directly injected in AppComponent which simplifies the code and makes the usage of `FormBuilder` direct and clear. This change also ensures a better use of Javascript's destructuring feature.