Fix typo and improve explanation in README
This commit corrects a typo in the terminology of `formCpntrolName` to `formControlName` for accuracy. Also, it has clarified that the child component can define a form control without needing to pass the form group to the child component. The change provides a better comprehension of the role and function of the child component in the form group.