Reformat HTML layout and update application title
The commit addresses two primary changes. First, it reorganizes the form inputs order and the `div` tag containers in the `app.component.html` to improve the app's layout. Especially, it enhances the presentation of the 'entered values' section to ensure easier readability. Second, the application's title has been updated in `index.html` from 'FormTest' to 'Angular Forms' for better context. The changes aim to enhance user experience and readability.